# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZIP 042 API",
    "version": "rules-2026-08-20.1 / extraction-2026-08-20",
    "description": "사회초년생을 위한 원룸 계약 안전 검사 서비스 백엔드.\n\n## 인증\n`Authorization: Bearer <Supabase access token>`\n목 모드(`ZIP042_MODE=mock`)에서는 아무 문자열이나 보내면 되고, 토큰이 곧 사용자 식별자입니다.\n\n## 시작하기\n1. `GET /v1/meta` — 기능 가용성 확인 (시세 조회 가능 여부 등)\n2. `GET /v1/addresses/search` — 주소를 골라 좌표·법정동코드를 얻습니다 (**필수**)\n3. `POST /v1/cases` — 검사 건 생성\n4. `POST /v1/cases/{caseId}/documents/upload-url` → PUT → `POST .../documents`\n5. `POST /v1/cases/{caseId}/analyze/jobs` — 비동기 분석 (권장) 또는 `POST .../analyze` (동기)\n\n## 판정의 두 축\n`verdict.score` 는 **실제 위험**만 반영합니다. 확인하지 못한 항목은 `informationGaps` 로\n따로 나가고, 그중 판정을 완료할 수 없게 만드는 것은 `blockingGaps` 입니다.\n`blockingGaps` 가 비어 있지 않으면 `contractable` 은 false 이며, 이때 UI 는\n\"위험합니다\"가 아니라 **\"아직 판단할 수 없습니다\"** 로 표현해야 합니다.\n\n## 주의\n모든 판정 응답에는 `caveats` / `disclaimer` 가 포함됩니다. **법률 자문이 아니라는 고지를**\n**UI 에 반드시 노출해야 합니다.**"
  },
  "servers": [
    {
      "url": "http://localhost:8787"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "UNPROCESSABLE",
                  "VALIDATION_FAILED",
                  "RATE_LIMITED",
                  "UPSTREAM_FAILED",
                  "INTERNAL"
                ]
              },
              "message": {
                "type": "string",
                "description": "사용자에게 그대로 보여줄 수 있는 한국어 메시지"
              },
              "issues": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "Case": {
        "type": "object",
        "required": [
          "id",
          "status",
          "property",
          "terms",
          "schedule"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "ready",
              "analyzing",
              "analyzed",
              "failed"
            ]
          },
          "property": {
            "type": "object",
            "properties": {
              "roadAddress": {
                "type": "string",
                "nullable": true
              },
              "jibunAddress": {
                "type": "string",
                "nullable": true
              },
              "detailAddress": {
                "type": "string",
                "nullable": true
              },
              "regionCode": {
                "type": "string",
                "nullable": true,
                "description": "법정동코드. 앞 5자리가 시군구코드"
              },
              "sigungu": {
                "type": "string",
                "nullable": true
              },
              "lat": {
                "type": "number",
                "nullable": true
              },
              "lng": {
                "type": "number",
                "nullable": true
              },
              "buildingType": {
                "type": "string",
                "nullable": true
              },
              "exclusiveAreaM2": {
                "type": "number",
                "nullable": true
              },
              "floor": {
                "type": "integer",
                "nullable": true
              },
              "totalFloors": {
                "type": "integer",
                "nullable": true
              },
              "builtYear": {
                "type": "integer",
                "nullable": true
              },
              "householdCount": {
                "type": "integer",
                "nullable": true
              }
            }
          },
          "terms": {
            "type": "object",
            "properties": {
              "leaseType": {
                "type": "string",
                "enum": [
                  "jeonse",
                  "monthly",
                  "semi_jeonse"
                ]
              },
              "depositKrw": {
                "type": "integer",
                "description": "원 단위"
              },
              "monthlyRentKrw": {
                "type": "integer"
              },
              "maintenanceFeeKrw": {
                "type": "integer"
              },
              "contractTermMonths": {
                "type": "integer"
              },
              "userMarketPriceKrw": {
                "type": "integer",
                "nullable": true
              }
            }
          },
          "schedule": {
            "type": "object",
            "properties": {
              "contractDate": {
                "type": "string",
                "format": "date",
                "nullable": true
              },
              "balanceDate": {
                "type": "string",
                "format": "date",
                "nullable": true
              },
              "moveInDate": {
                "type": "string",
                "format": "date",
                "nullable": true
              },
              "residentRegistrationDate": {
                "type": "string",
                "format": "date",
                "nullable": true
              },
              "confirmedDatePlan": {
                "type": "string",
                "format": "date",
                "nullable": true
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateCaseRequest": {
        "type": "object",
        "properties": {
          "amountUnit": {
            "type": "string",
            "enum": [
              "krw",
              "man"
            ],
            "default": "krw"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120,
            "nullable": true
          },
          "roadAddress": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300,
            "nullable": true
          },
          "jibunAddress": {
            "type": "string",
            "maxLength": 300,
            "nullable": true
          },
          "detailAddress": {
            "type": "string",
            "maxLength": 100,
            "nullable": true
          },
          "regionCode": {
            "type": "string",
            "pattern": "^\\d{5,10}$",
            "nullable": true
          },
          "sigungu": {
            "type": "string",
            "maxLength": 60,
            "nullable": true
          },
          "lat": {
            "type": "number",
            "minimum": -90,
            "maximum": 90,
            "nullable": true
          },
          "lng": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "nullable": true
          },
          "buildingType": {
            "type": "string",
            "enum": [
              "apartment",
              "officetel",
              "multi_family",
              "multi_household",
              "row_house",
              "detached",
              "studio",
              "other"
            ],
            "nullable": true
          },
          "exclusiveAreaM2": {
            "type": "number",
            "exclusiveMinimum": true,
            "minimum": 0,
            "maximum": 10000,
            "nullable": true
          },
          "floor": {
            "type": "integer",
            "minimum": -10,
            "maximum": 200,
            "nullable": true
          },
          "totalFloors": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200,
            "nullable": true
          },
          "builtYear": {
            "type": "integer",
            "minimum": 1900,
            "maximum": 2100,
            "nullable": true
          },
          "householdCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000,
            "nullable": true
          },
          "leaseType": {
            "type": "string",
            "enum": [
              "jeonse",
              "monthly",
              "semi_jeonse"
            ],
            "default": "monthly"
          },
          "deposit": {
            "type": "integer",
            "minimum": 0,
            "default": 0
          },
          "monthlyRent": {
            "type": "integer",
            "minimum": 0,
            "default": 0
          },
          "maintenanceFee": {
            "type": "integer",
            "minimum": 0,
            "default": 0
          },
          "userMarketPrice": {
            "type": "integer",
            "minimum": 0,
            "nullable": true
          },
          "contractTermMonths": {
            "type": "integer",
            "minimum": 1,
            "maximum": 120,
            "default": 24
          },
          "contractDate": {
            "type": "string",
            "nullable": true
          },
          "balanceDate": {
            "type": "string",
            "nullable": true
          },
          "moveInDate": {
            "type": "string",
            "nullable": true
          },
          "residentRegistrationDate": {
            "type": "string",
            "nullable": true
          },
          "confirmedDatePlan": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateScheduleRequest": {
        "type": "object",
        "properties": {
          "contractDate": {
            "type": "string",
            "nullable": true
          },
          "balanceDate": {
            "type": "string",
            "nullable": true
          },
          "moveInDate": {
            "type": "string",
            "nullable": true
          },
          "residentRegistrationDate": {
            "type": "string",
            "nullable": true
          },
          "confirmedDatePlan": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MarketPriceOverrideRequest": {
        "type": "object",
        "properties": {
          "amountUnit": {
            "type": "string",
            "enum": [
              "krw",
              "man"
            ],
            "default": "krw"
          },
          "estimated": {
            "type": "integer",
            "exclusiveMinimum": true,
            "minimum": 0
          },
          "note": {
            "type": "string",
            "maxLength": 200,
            "nullable": true
          }
        },
        "required": [
          "estimated"
        ],
        "additionalProperties": false
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/cases": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "검사 건 목록",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cases": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Case"
                      }
                    },
                    "pagination": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "검사 건 생성",
        "description": "금액은 `amountUnit` 으로 단위를 명시합니다(`krw` 또는 `man`). 좌표·법정동코드를 비워두면 서버가 주소로 보완을 시도합니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "case": {
                      "$ref": "#/components/schemas/Case"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{caseId}": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "검사 건 상세",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "상세",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "case": {
                      "$ref": "#/components/schemas/Case"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "cases"
        ],
        "summary": "검사 건 수정",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cases"
        ],
        "summary": "검사 건 삭제",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "삭제됨"
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{caseId}/schedule": {
      "patch": {
        "tags": [
          "schedule"
        ],
        "summary": "일정만 수정",
        "description": "일정이 바뀌면 대항력 발생 시점이 바뀌므로 재분석이 필요합니다.",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{caseId}/market-price": {
      "put": {
        "tags": [
          "cases"
        ],
        "summary": "시세 직접 입력",
        "description": "공공 API 조회가 실패했을 때의 폴백 경로입니다.",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketPriceOverrideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "저장됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "입력값 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "대상 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "요청 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}